Fix reading of negative temperatures (Fixes #24)

+++ Adafruit_SHT31.cpp
@@ -135,7 +135,7 @@ bool Adafruit_SHT31::readTempHum(void) {
       readbuffer[5] != crc8(readbuffer + 3, 2))
     return false;
 
-  uint32_t stemp = ((uint32_t)readbuffer[0] << 8) | readbuffer[1];
+  int32_t stemp = (int32_t) (((uint32_t)readbuffer[0] << 8) | readbuffer[1]);
   // simplified (65536 instead of 65535) integer version of:
   //temp = (stemp * 175.0f) / 65535.0f - 45.0f;
   stemp = ((4375 * stemp) >> 14) - 4500;
